Add buttons for game UI
Allows user to play the game through buttons rather than the
console, which is more seamless and easier, and prevents bad
input from the part of the user. Attaching event listeners
allows us to listen to the user clicks to observe what choice
of buttons they made, and then, the game is played from that
choice.