Add logic to display score and winner
We added the logic required to keep track of the score of the
player and the computer, and then we used it to determine a
winner depending on who reached a score of 5 first. All of the
changes made, were made to reach this goal of being able to play
the game from the UI using the buttons.